engine: honest wasm-backend reject for conveyor/queue models
The wasm code-generation backend does not lower the conveyor/queue
side-table passes; such models were caught only by the NotExpanded
guard, whose error told a wasm caller to "go through
queue_compile::build_vm" -- a VM-only path that produces no wasm blob.

compile_datamodel_to_artifact now rejects up front with a typed
WasmGenError::Unsupported stating the real limitation, using the same
project_has_conveyor/project_has_queue predicates every simulation
surface dispatches on (shared main_model_has_stock core, so the wasm
reject and the VM dispatch cannot disagree). The db.rs guard stays as
the path-neutral backstop, reworded to name the bytecode VM as the
only backend for these constructs. Engine and libsimlin FFI tests pin
the typed error (no panic, NULL output buffers); the AC3.2 parity
hook's implicit conveyor-fixture exclusion is now documented at the
hook; both CLAUDE.md wasmgen sections list conveyor/queue as the third
unlowerable class. Full lowering remains tracked in GH #884.

module.rs crossed the per-file line cap with the new test, so its
test module moves to the sibling module_tests.rs via the same #[path]
mount lower_tests.rs uses.
