+++ src/libsimlin/CLAUDE.md
@@ -41,7 +41,7 @@ All public FFI functions are prefixed with `simlin_` and declared `extern "C"`. 
 - **`src/model.rs`** - Inspect model structure:
   - `simlin_model_{ref,unref}()`, `simlin_model_get_var_count()`, `simlin_model_get_var_names()`
   - `simlin_model_get_dependencies()`, `simlin_model_get_links()`, `simlin_model_get_equations()`
-  - `simlin_model_compile_to_wasm(model, ltm_enabled, ltm_discovery_mode, out_wasm, out_wasm_len, out_layout, out_layout_len, out_error)` - Compile the model to a self-contained wasm module (engine `wasmgen` backend, an alternative to the VM for fast repeated re-simulation). Returns two malloc'd buffers, each freed with `simlin_free`: the wasm blob and a serialized `WasmLayout` (length-prefixed, little-endian: geometry `n_slots`/`n_chunks`/`results_offset` then a canonical-name -> slot-offset map a host strides the results region with). Works from a `SimlinModel`'s datamodel alone -- no `SimlinSim` required -- and stores a `SimlinError` (never panics) on any compile/codegen failure. The two bool flags thread through to the salsa compile so `ltm_enabled` produces a blob whose results region carries the LTM link/loop-score columns the VM emits (read back via the from-wasm analyze FFIs in `src/analysis.rs`); `ltm_discovery_mode` additionally enables the discovery-mode synthetic variables. An unlowerable LTM construct still surfaces a `SimlinError` -- no silent VM fallback
+  - `simlin_model_compile_to_wasm(model, ltm_enabled, ltm_discovery_mode, out_wasm, out_wasm_len, out_layout, out_layout_len, out_error)` - Compile the model to a self-contained wasm module (engine `wasmgen` backend, an alternative to the VM for fast repeated re-simulation). Returns two malloc'd buffers, each freed with `simlin_free`: the wasm blob and a serialized `WasmLayout` (length-prefixed, little-endian: geometry `n_slots`/`n_chunks`/`results_offset` then a canonical-name -> slot-offset map a host strides the results region with). Works from a `SimlinModel`'s datamodel alone -- no `SimlinSim` required -- and stores a `SimlinError` (never panics) on any compile/codegen failure. The two bool flags thread through to the salsa compile so `ltm_enabled` produces a blob whose results region carries the LTM link/loop-score columns the VM emits (read back via the from-wasm analyze FFIs in `src/analysis.rs`); `ltm_discovery_mode` additionally enables the discovery-mode synthetic variables. An unlowerable LTM construct still surfaces a `SimlinError` -- no silent VM fallback. A conveyor/queue model is rejected up front with a clear `SimlinError` naming the wasm-backend limitation (GH #884: the conveyor/queue side-table passes are implemented only in the bytecode VM, so `simlin_sim_new`'s VM path is the only way to simulate such a model)
 
 ### Standalone results (VDF import)
 
@@ -92,7 +92,7 @@ Integration tests live in the single consolidated `tests/integration` harness (o
 - **`tests/integration/patch.rs`** - JSON patch application, error collection, unit warnings, XMILE patches
 - **`tests/integration/incremental.rs`** - Incremental compilation path (patch-then-sim, snapshot isolation)
 - **`tests/integration/analysis.rs`** - Causal analysis: incoming links, loop detection, loop scores
-- **`tests/integration/wasm.rs`** - `simlin_model_compile_to_wasm`: validates and executes the returned blob under the DLR-FT interpreter (a libsimlin dev-dependency), parses the returned layout per its documented wire format, and checks the strided series against the VM via `simlin_sim_get_series`; also asserts a graceful `SimlinError` (no panic) for an unsupported model. `compile_to_wasm_blob_supports_resumable_run` further drives the FFI-compiled blob's resumable ABI (`run_initials`/`run_to`/`set_value`/`reset`, the additive exports) against a VM oracle driven identically through `simlin_sim_run_to`/`simlin_sim_set_value`/`simlin_sim_run_to_end`, confirming the resumable surface survives the FFI compile path and the export set grew purely additively. `compile_to_wasm_unsupported_ltm_model_surfaces_error` exercises the `ltm_enabled=true` path against an intentionally unlowerable LTM model and asserts the `SimlinError` reaches the caller (no silent VM fallback). The from-wasm analysis twins are covered by `links_from_wasm_match_vm` and `rel_loop_score_from_wasm_matches_vm`: each compiles an LTM-enabled wasm blob, runs it under the interpreter to extract the slab, and asserts that `simlin_analyze_links_from_wasm_results` / `simlin_analyze_rel_loop_score_from_wasm_results` agree with their VM-side FFI counterparts on the link set and per-step scores
+- **`tests/integration/wasm.rs`** - `simlin_model_compile_to_wasm`: validates and executes the returned blob under the DLR-FT interpreter (a libsimlin dev-dependency), parses the returned layout per its documented wire format, and checks the strided series against the VM via `simlin_sim_get_series`; also asserts a graceful `SimlinError` (no panic) for an unsupported model. `compile_to_wasm_blob_supports_resumable_run` further drives the FFI-compiled blob's resumable ABI (`run_initials`/`run_to`/`set_value`/`reset`, the additive exports) against a VM oracle driven identically through `simlin_sim_run_to`/`simlin_sim_set_value`/`simlin_sim_run_to_end`, confirming the resumable surface survives the FFI compile path and the export set grew purely additively. `compile_to_wasm_unsupported_ltm_model_surfaces_error` exercises the `ltm_enabled=true` path against an intentionally unlowerable LTM model and asserts the `SimlinError` reaches the caller (no silent VM fallback). `compile_to_wasm_conveyor_and_queue_models_surface_error` pins the GH #884 up-front reject: a conveyor/queue model surfaces a typed error stating the wasm-backend limitation (never the engine-internal NotExpanded guard text), with both output buffers NULL. The from-wasm analysis twins are covered by `links_from_wasm_match_vm` and `rel_loop_score_from_wasm_matches_vm`: each compiles an LTM-enabled wasm blob, runs it under the interpreter to extract the slab, and asserts that `simlin_analyze_links_from_wasm_results` / `simlin_analyze_rel_loop_score_from_wasm_results` agree with their VM-side FFI counterparts on the link set and per-step scores
 - **`tests/integration/rendering.rs`** - SVG and PNG diagram rendering
 - **`tests/integration/results.rs`** - Standalone results FFI via VDF import: run-file and dataset-file opens with pinned series values (fixtures under `/test/bobby/vdf/`), canonical + display-name series lookup, sorted name listing, malformed-input rejection (empty/garbage/truncated/NULL) without crashing, refcounting, and NULL-handle errors
 - **`tests/integration/diagram.rs`** - Diagram layout sync
